==> scripts/5_wgcna_tensor_decomp.ipynb <==
### Nonnegative tensor decomposition of co-expression networks (TOMs)

# libraries
import os
import numpy as np
import pandas as pd
import tensortools as tt

# get parent wd
os.chdir('..')
PATH = os.getcwd()

tom_tensor = []

for filename in os.listdir(PATH+'/data'):
    if "tom_mrna" in filename:
        print(filename)
        df = pd.read_csv(
            PATH+'/data/'+filename,
            index_col =None,
            header = 0,
            sep = ",",
            compression = 'gzip'
        )
        tmp = df.values
        tom_tensor.append(tmp)

# assemble tensor
tom_tensor = np.asarray(tom_tensor)
tom_tensor = np.swapaxes(tom_tensor,0,2)
tom_tensor = tom_tensor.astype('float32')

# get gene names
names = df.columns.values

# print dims
print(tom_tensor.shape)

# number of factors
R=10

# perform tensor decomposition
U = tt.ncp_hals(
    tom_tensor,
    rank=R,
    verbose=False
)
# get factors
factors_tt = U.factors.factors


## convert matrices to data frames
A = pd.DataFrame(factors_tt[0], index = names.tolist())
B = pd.DataFrame(factors_tt[1], index = names.tolist())
C = pd.DataFrame(factors_tt[2])

# write the output
pd.DataFrame.to_csv(
    A, PATH+'/results/rna_A_factor.csv',
    header=None
)
pd.DataFrame.to_csv(
    B, PATH+'/results/rna_B_factor.csv',
    header=None
)
pd.DataFrame.to_csv(
    C, PATH+'/results/rna_C_factor.csv',
    header=None
)
