==> scripts/6_component_annotation.Rmd <==
---
title: "Functional annotation of co-expression components"
output: rmarkdown::github_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

```

Load prerequisites that will be needed for gene set enrichment analysis (GSEA) in R environment. NOTE: here we use conda environment, which contains [kneed](https://kneed.readthedocs.io/en/stable/) python package. The package will be used to obtain lead driving genes of each co-expression component.
```{r, message=FALSE, warning=FALSE}

# libraries
library(tidyverse)
library(clusterProfiler)
library(org.Hs.eg.db)
library(reticulate)

## define Python, which contain kneed
## please change this depending on your env
use_condaenv(
  condaenv = "reticulate_env",
  required = TRUE
  )
# import kneed
kneed <- import("kneed")

```

Read results of nonnegative CP tensor decomposition, which were generated using 5_wgcna_tensor_decomp.ipynb:
```{r, warning=FALSE, message=FALSE}

# set wd to main
setwd('..')

# components results
component_list <- 
  list.files(
    path = paste0(getwd(), "/results"),
    pattern = "factor.csv",
    full.names = TRUE
  )
names(component_list) <- str_extract(component_list, "[A-C]_factor")

# read
component_list <- lapply(
  component_list,
  read_csv,
  col_names=c(
    "var", paste0("comp#",1:10)
  )
  )

```

Knee point determination to identify the lead driving genes:
```{r, warning=FALSE, message=FALSE}

# use scores of A latent matrix
A <- component_list$A_factor %>%
  gather(comp, score, -var) %>% 
  split(.$comp)

# sort and add order
A <- lapply(A, function(j){
  j %>% arrange(desc(score)) %>% 
    mutate(order = rep(1:n()))
  })

# calculate knee points
A_knee <- lapply(A, function(j){
  kneed$KneeLocator(
    x= j[,4, drop=T],
    y= j[,3, drop=T],
    S = 1,
    curve='convex',
    direction='decreasing'
    ) %>% .$knee
  })

```

Functional annotation of co-expression components using GSEA:
```{r, warning=FALSE, message=FALSE}

# get gene lists
component_knee_genes <- 
  lapply(seq(A), function(i){
    A[[i]] %>% filter(A_knee[[i]] >= order) %>% 
      pull(var) %>%
      bitr(.,
           fromType="SYMBOL",
           toType="ENTREZID",
           OrgDb = org.Hs.eg.db,
           drop = TRUE
           )
})
names(component_knee_genes) <- names(A)

# run enrichment analysis
component_knee_go <- 
  lapply(component_knee_genes, function(x){
    enrichGO(
      gene = x$ENTREZID,
      OrgDb = org.Hs.eg.db,
      ont = "BP",
      pAdjustMethod = "BH", 
      pvalueCutoff  = 0.05,
      qvalueCutoff  = 0.05
      )
    })

# flatten results to table
component_knee_go_tb <- lapply(
  component_knee_go, function(x) x %>% .@result
  ) %>% bind_rows(.id = "component") %>%
  filter(p.adjust<0.05) %>%
  group_by(component) %>%
  top_n(-5, wt=p.adjust)

# visualize
component_knee_go_tb %>% 
  ggplot(aes(x = Count, y = fct_reorder(Description, Count))) + 
  geom_point(aes(size = Count, color = p.adjust)) +
  facet_wrap(~component, nrow = 1) + ylab(NULL) +
  theme(legend.position = "bottom",
        legend.direction = "horizontal",
        legend.box = "vertical")

```
```{r, warning=FALSE, message=FALSE}
sessionInfo()
```
